Move autosave package.
