--- src/main/java/net/sf/jabref/autosave/event/AutosaveEvent.java
@@ -1,10 +1,0 @@
-package net.sf.jabref.autosave.event;
-
-import net.sf.jabref.autosave.AutosaveManager;
-
-/**
- * This Event is fired from {@link AutosaveManager} in case that a save task is pending.
- */
-public class AutosaveEvent {
-    // no data
-}
