+++ src/main/java/net/sf/jabref/gui/autosave/AutosaveUIManager.java
@@ -1,8 +1,8 @@
 package net.sf.jabref.gui.autosave;
 
-import net.sf.jabref.autosave.event.AutosaveEvent;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.exporter.SaveDatabaseAction;
+import net.sf.jabref.model.database.event.AutosaveEvent;
 
 import com.google.common.eventbus.Subscribe;
 import org.apache.commons.logging.Log;
